refactored news feed for better components and adjusted code again to use newscatcher API
