--- src/components/NewsComponent/newsComponent.module.css
@@ -1,48 +1,0 @@
-.search_bubbles {
-  display: flex;
-  justify-content: space-around;
-  margin: 16px 0;
-}
-
-.bubble img {
-  width: 40px;
-  display: inline;
-}
-
-.search {
-  margin: 8px auto;
-  display: flex;
-  justify-content: center;
-}
-.newsfeed {
-  width: 90%;
-  margin: 16px auto;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-}
-
-.news_images {
-  margin: 0 auto;
-  width: 100%;
-}
-
-.news_excerpt {
-  line-height: 1.5em;
-  color: var(--primary-accent);
-}
-.api_space {
-  height: 20vh;
-  font-weight: bold;
-  background-color: paleturquoise;
-  width: 100%;
-  text-align: center;
-  padding-top: 20px;
-}
-
-.news_item {
-  margin: 16px 0;
-  color: inherit;
-  text-decoration: inherit;
-}
