+++ src/components/newsSection/newsSection.module.css
@@ -1,0 +1,20 @@
+.news_section {
+}
+
+.newsfeed {
+  width: 90%;
+  margin: 16px auto;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+}
+
+.api_space {
+  height: 20vh;
+  font-weight: bold;
+  background-color: paleturquoise;
+  width: 100%;
+  text-align: center;
+  padding-top: 20px;
+}
