+++ src/components/searchBubbles/searchBubbles.module.css
@@ -1,0 +1,10 @@
+.search_bubbles {
+  display: flex;
+  justify-content: space-around;
+  margin: 16px 0;
+}
+
+.bubble img {
+  width: 40px;
+  display: inline;
+}
