+++ src/components/searchComponent/searchComponent.module.css
@@ -1,0 +1,5 @@
+.search {
+  margin: 8px auto;
+  display: flex;
+  justify-content: center;
+}
